fix: update data types for nutrient values to float in multiple models

+++ models/crop_rec/app.py
@@ -14,9 +14,9 @@ model = joblib.load(os.path.join(ARTIFACTS_DIR, "model.pkl"))
 scaler = joblib.load(os.path.join(ARTIFACTS_DIR, "scaler.pkl"))
 
 class Items(BaseModel):
-    N: int
-    P: int
-    K: int
+    N: float
+    P: float
+    K: float
     temperature: float
     humidity: float
     ph: float
